[KDA] add SOLVE_TRIL_DOT_PRECISION for matrix inverse precision in chunk_intra
- Use tf32x3 precision for tl.dot in solve_tril to improve matrix inverse accuracy
- Cast loaded g values to float32 before exp() operations
- Clean up chunk_kda_fwd_intra function signature (remove unused params)

+++ fla/ops/kda/chunk.py
@@ -36,8 +36,7 @@ def chunk_kda_fwd(
         beta=beta,
         scale=scale,
         cu_seqlens=cu_seqlens,
-        chunk_indices=chunk_indices,
-        output_dtype=torch.float32,
+        chunk_indices=chunk_indices
     )
     w, u, _, kg = recompute_w_u_fwd(
         k=k,
